Extract configuration options to Configuration
Right now, configuration options are class methods on SpreeSocial. These
options are hard to test because if you want to set an option in a test,
then it changes global state and you have to make sure to roll it back
so that it doesn't affect future tests. Instead of doing that, extract
these options to a class, so that when setting an option you will do so
on a Configuration instance.