Redirect to root path on auth failure
When a user is logging into Spree for the first time, they will get an
screen that asks if they want to authorize a connection to the OAuth
application.

If the user rejects this authorization, then by default, they will be
redirected to the Spree login page. The problem is, wu-unit overrides
this route so that the user is automatically sent to Gatekeeper instead.

We don't really need to go to that page anymore, so simply redirect to
the root path instead.