Fix OmniAuth auth failure proc
Currently we configure OmniAuth so that if the authorization step fails,
the `failure` action in Spree::OmniauthCallbacksController is called.
This controller is obtained by taking the configured Spree user class
(Spree::User), asking for its table name (`spree_users`), then looking
up the singular version of that table name (`:spree_user`) in the
`Devise.mappings` hash.

The problem is, since we've reconfigured Postgres to use multiple
tenants, the result of asking for the table name of any model is going
to start with `public.` (or whatever the tenant is, plus `.`).
So the key we look up might be `:'public.spree_user'`, not
`:spree_user`.

Fix this by essentially taking away the tenant name before looking it up
in `Devise.mappings`. Also, use `fetch` to look up the key, so if it
doesn't exist, we get a more informative error immediately.